Work around wine compatibility issue.
Triggers were failing to countdown on wine. After investigation I determined it was because there's a bug in Wine's WPF text rendering system that causes the MobInfo window to fail to load. Triggers were loaded after mobinfo so the application never got to that state.

The "fix" here is to gracefully skip any windows that fail to load. This allows the triggers to work even though mobinfo fails. I can imagine this getting fixed in a later version of wine, so it's possible it might start magically working in the future.

In the meantime: known issue MobInfo fails to load in linux due to a bug _probably_ in wine.